Create UmpleVersion3
Fixed the assumption that Java Drawing was to be used not System.out. Code now draws under the assumption that only the Board class will be printing a UI. Added methods to hide and show the grid. Added a hand printer for Player and added a basic framework to all Item childs and Item itself